parser: introduce global_.. and fmt_parse_until()
the previous `parse_until()` function is renamed to
`global_parse_until()` and a new function `fmt_parse_until()`, which
parses a subset of the global one ist introduced.

this is to differentiate between different parsing contexts: in the
global parsing context, we can parse more constructs, than in the `fmt`
context, where we only parse _formatting_ constructs (hence the name).